+++ nlmModulePy.cpp
@@ -17,7 +17,7 @@ PYBIND11_MODULE(nonlinearmedium, m) {
 
   // default arguments for Python initialization of empty arrays
   Eigen::Ref<const Arraycd> defArraycd = Eigen::Ref<const Arraycd>(Arraycd{});
-  Eigen::Ref<const Arrayf>  defArrayf  = Eigen::Ref<const Arrayf>(Arrayf{});
+  Eigen::Ref<const Arrayd>  defArrayf  = Eigen::Ref<const Arrayd>(Arrayd{});
 
 /*
  * Chi3
@@ -65,7 +65,7 @@ PYBIND11_MODULE(nonlinearmedium, m) {
 
   Chi2PDC.def(
       py::init<double, double, double, double, double, Eigen::Ref<const Arraycd>&, int,
-               double, double, double, double, double, double, double, uint, uint, Eigen::Ref<const Arrayf>&>(),
+               double, double, double, double, double, double, double, uint, uint, Eigen::Ref<const Arrayd>&>(),
       "relativeLength"_a, "nlLength"_a, "dispLength"_a, "beta2"_a, "beta2s"_a, "customPump"_a = defArraycd, "pulseType"_a = 0,
       "beta1"_a = 0, "beta1s"_a = 0, "beta3"_a = 0, "beta3s"_a = 0, "diffBeta0"_a = 0,
       "chirp"_a = 0, "tMax"_a = 10, "tPrecision"_a = 512, "zPrecision"_a = 100, "poling"_a = defArrayf);
@@ -110,7 +110,7 @@ PYBIND11_MODULE(nonlinearmedium, m) {
   Chi2SFG.def(
       py::init<double, double, double, double, double, double, double, Eigen::Ref<const Arraycd>&, int,
                double, double, double, double, double, double, double, double, double, double, uint, uint,
-               Eigen::Ref<const Arrayf>&>(),
+               Eigen::Ref<const Arrayd>&>(),
       "relativeLength"_a, "nlLength"_a, "nlLengthOrig"_a, "dispLength"_a, "beta2"_a, "beta2s"_a, "beta2o"_a,
       "customPump"_a = defArraycd, "pulseType"_a = 0, "beta1"_a = 0, "beta1s"_a = 0, "beta1o"_a = 0, "beta3"_a = 0, "beta3s"_a = 0, "beta3o"_a = 0, "diffBeta0"_a = 0,
       "diffBeta0o"_a = 0, "chirp"_a = 0, "tMax"_a = 10, "tPrecision"_a = 512, "zPrecision"_a = 100, "poling"_a = defArrayf);
